+++ metadata.rb
@@ -19,6 +19,7 @@ depends          'apt', '~> 2.8'
 
 supports         'mac_os_x'
 supports         'windows'
-supports         'fedora'
 supports         'ubuntu'
+supports         'debian'
+supports         'fedora'
 # rubocop:enable SingleSpaceBeforeFirstArg